+++ src/routes/newsRoute.js
@@ -3,19 +3,16 @@ import { createNews, getAllNews, getNewsById, updateNews, deleteNews } from '../
 
 const router = express.Router();
 
-// Crear una noticia
+// Middleware para registrar las solicitudes
+router.use((req, res, next) => {
+    console.log(`Solicitud ${req.method} en la ruta ${req.originalUrl}`);
+    next(); // Llama a next() para pasar al siguiente middleware o controlador
+});
+
 router.post('/', createNews);
-
-// Obtener todas las noticias
 router.get('/', getAllNews);
-
-// Obtener una noticia por ID
 router.get('/:id', getNewsById);
-
-// Actualizar una noticia
 router.put('/:id', updateNews);
-
-// Eliminar una noticia
 router.delete('/:id', deleteNews);
 
 export default router;
