+++ src/routes/userRoutes.js
@@ -3,6 +3,12 @@ import { deleteUser, followUser, getUser, UnFollowUser, updateUser } from "../co
 
 const router = express.Router();
 
+// Middleware para registrar las solicitudes
+router.use((req, res, next) => {
+    console.log(`Solicitud ${req.method} en la ruta ${req.originalUrl}`);
+    next();
+});
+
 router.get('/:id', getUser)
 router.put('/:id', updateUser)
 router.delete('/:id', deleteUser)
